First attempt at ignoring commits on topic branch that are from a rebase from master.
When creating a commit email, ignore revisions on a topic branch that
have been already applied to master. This keeps the commit email cleaner
by only providing the commits that actually pertain ONLY to the topic branch.

This only affects the main body of the email, the summary section still needs
to be looked at, since that is just the raw output of git log. Any help
on refactoring that section would be most welcome :)

I haven't updated dev.horde.org yet to this new script. Would appreciate someone else
looking at it to make sure it looks sane first.