First attempt at filtering the commit summary by commits that exist on master
Perform similar filtering for the commit summary section of the email. Probably
a more efficient way to do this...but this is a start. Again, not updating dev.h.o
until some other eyes can sanity check this. Also, we should probably capture the
commit ids of those we are silently ignoring and provide *some* feedback that the
branch was sync'd.