Revert "This isn't needed anymore - CI does this for us"
This reverts commit 9c2d867a3e1ec55bb68ece480af58f6348beddf5.

I think it makes sense to keep the Horde specific CodeSniffer ruleset.
For one thing as a reference to see what we actually use on our code.
The CI hides this too much as the ruleset is hidden in the php-ci-tools
repo and not that visible. And people can include this ruleset locally
within their IDE to directly check for problems while coding.

In the long run I'll probably integrate the ruleset with the components
package. But for now I keep it here.

An upgrade for the newest CodeSniffer is pending which is why I revive
this directory now.