Fix PEAR install issue
Unfortunately, the way we build Horde_Translation creates static paths
in the file, which totally breaks when used in the phar environment.

For now, do a search/replace directly on these files to fix the issue.

Performance improvement: keep local copy of composer.phar so we don't
need to download when running install build script multiple times